fix: css file rename
